Show dashboard on root page for authenticated users
- Root page (/) now shows dashboard content for logged-in users
- Unauthenticated users still see the hero splash page
- Removed Dashboard tab from navbar since / serves that purpose
- Extracted DashboardContent to reusable component
