+++ src/app/page.tsx
@@ -1,7 +1,28 @@
+'use client';
+
 import Link from 'next/link';
 import { Button } from '@/components/ui';
+import { useAuth } from '@/contexts/AuthContext';
+import { DashboardContent } from '@/components/dashboard/DashboardContent';
 
 export default function Home() {
+  const { profile, isLoading } = useAuth();
+
+  // Show dashboard for authenticated users
+  if (profile) {
+    return <DashboardContent />;
+  }
+
+  // Show loading state while checking auth
+  if (isLoading) {
+    return (
+      <div className="min-h-[calc(100vh-4rem)] flex items-center justify-center">
+        <div className="h-8 w-8 animate-spin rounded-full border-4 border-primary border-t-transparent" />
+      </div>
+    );
+  }
+
+  // Show hero page for unauthenticated users
   return (
     <div className="min-h-[calc(100vh-4rem)]">
       {/* Hero Section */}
