+++ src/components/dashboard/DashboardContent.tsx
@@ -1,0 +1,178 @@
+'use client';
+
+import { useAuth } from '@/contexts/AuthContext';
+import { useEffect, useState } from 'react';
+import Link from 'next/link';
+import { Button, Card, CardContent, EmptyState, LoadingSpinner } from '@/components/ui';
+import { createClient } from '@/lib/supabase/client';
+import { GameWithGM } from '@/types';
+import { DAY_LABELS } from '@/lib/constants';
+
+interface GameWithGMAndCount extends GameWithGM {
+  member_count: number;
+}
+
+export function DashboardContent() {
+  const { profile, isLoading, session } = useAuth();
+  const [games, setGames] = useState<GameWithGMAndCount[]>([]);
+  const [loading, setLoading] = useState(true);
+  const supabase = createClient();
+
+  useEffect(() => {
+    async function fetchGames() {
+      if (!profile?.id) {
+        // No profile, nothing to fetch
+        setLoading(false);
+        return;
+      }
+
+      // Fetch games where user is GM or member
+      const { data: memberships } = await supabase
+        .from('game_memberships')
+        .select('game_id')
+        .eq('user_id', profile.id);
+
+      const memberGameIds = memberships?.map((m) => m.game_id) || [];
+
+      const { data: gmGames } = await supabase
+        .from('games')
+        .select('*, gm:users!games_gm_id_fkey(*)')
+        .eq('gm_id', profile.id);
+
+      const { data: memberGames } = await supabase
+        .from('games')
+        .select('*, gm:users!games_gm_id_fkey(*)')
+        .in('id', memberGameIds.length > 0 ? memberGameIds : ['none']);
+
+      // Combine and dedupe
+      const allGames = [...(gmGames || []), ...(memberGames || [])];
+      const uniqueGames = Array.from(new Map(allGames.map((g) => [g.id, g])).values());
+
+      // Get member counts
+      const gamesWithCounts = await Promise.all(
+        uniqueGames.map(async (game) => {
+          const { count } = await supabase
+            .from('game_memberships')
+            .select('*', { count: 'exact', head: true })
+            .eq('game_id', game.id);
+
+          return { ...game, member_count: (count || 0) + 1 }; // +1 for GM
+        })
+      );
+
+      setGames(gamesWithCounts as GameWithGMAndCount[]);
+      setLoading(false);
+    }
+
+    if (profile?.id) {
+      fetchGames();
+    } else if (!isLoading) {
+      // Auth finished but no profile - stop loading
+      setLoading(false);
+    }
+  }, [profile?.id, isLoading]);
+
+  // Show spinner while auth is loading, data is loading, or profile hasn't loaded yet
+  if (isLoading || loading || (session && !profile)) {
+    return (
+      <div className="min-h-[calc(100vh-4rem)] flex items-center justify-center">
+        <LoadingSpinner />
+      </div>
+    );
+  }
+
+  return (
+    <div className="max-w-7xl mx-auto px-4 sm:px-6 lg:px-8 py-8">
+      <div className="flex justify-between items-center mb-8">
+        <div>
+          <h1 className="text-2xl font-bold text-foreground">Your Games</h1>
+          <p className="text-muted-foreground mt-1">Manage your campaigns and game sessions</p>
+        </div>
+        {profile?.is_gm && (
+          <Link href="/games/new">
+            <Button>Create New Game</Button>
+          </Link>
+        )}
+      </div>
+
+      {games.length === 0 ? (
+        <Card>
+          <CardContent className="py-12">
+            <EmptyState
+              icon="🎲"
+              title="No games yet"
+              description={
+                profile?.is_gm
+                  ? "Create your first game to start scheduling sessions with your group."
+                  : "Join a game using an invite link from your GM, or request GM status in settings to create your own."
+              }
+              action={
+                profile?.is_gm ? (
+                  <Link href="/games/new">
+                    <Button>Create Your First Game</Button>
+                  </Link>
+                ) : (
+                  <Link href="/settings">
+                    <Button variant="secondary">Go to Settings</Button>
+                  </Link>
+                )
+              }
+            />
+          </CardContent>
+        </Card>
+      ) : (
+        <div className="grid gap-6 md:grid-cols-2 lg:grid-cols-3">
+          {games.map((game) => (
+            <Link key={game.id} href={`/games/${game.id}`}>
+              <Card className="hover:shadow-md transition-shadow cursor-pointer h-full">
+                <CardContent className="py-6">
+                  <div className="flex items-start justify-between mb-4">
+                    <div>
+                      <h3 className="text-lg font-semibold text-card-foreground">{game.name}</h3>
+                      <p className="text-sm text-muted-foreground">
+                        GM: {game.gm.name}
+                        {game.gm_id === profile?.id && ' (You)'}
+                      </p>
+                    </div>
+                    {game.gm_id === profile?.id && (
+                      <span className="px-2 py-1 text-xs font-medium bg-primary/10 text-primary rounded">
+                        GM
+                      </span>
+                    )}
+                  </div>
+
+                  {game.description && (
+                    <p className="text-sm text-muted-foreground mb-4 line-clamp-2">{game.description}</p>
+                  )}
+
+                  <div className="flex items-center gap-4 text-sm text-muted-foreground">
+                    <span className="flex items-center gap-1">
+                      <span>👥</span>
+                      {game.member_count} player{game.member_count !== 1 ? 's' : ''}
+                    </span>
+                    <span className="flex items-center gap-1">
+                      <span>📅</span>
+                      {game.play_days.map((d) => DAY_LABELS.short[d]).join(', ')}
+                    </span>
+                  </div>
+                </CardContent>
+              </Card>
+            </Link>
+          ))}
+        </div>
+      )}
+
+      {!profile?.is_gm && (
+        <div className="mt-8 p-4 bg-amber-500/10 border border-amber-500/20 rounded-lg">
+          <p className="text-amber-700 dark:text-amber-400 text-sm">
+            <strong>Want to create your own games?</strong> Go to{' '}
+            <Link href="/settings" className="underline">
+              Settings
+            </Link>{' '}
+            to enable GM mode.
+          </p>
+        </div>
+      )}
+    </div>
+  );
+}
